Divide classes into packages
