+++ src/main/java/reo/Event.java
@@ -1,3 +1,5 @@
+package reo;
+
 public class Event extends Task {
     private String from;
     private String to;
