+++ src/main/java/reo/Task.java
@@ -1,3 +1,5 @@
+package reo;
+
 public class Task {
     private boolean isDone;
     private String name;
