+++ text-ui-test/EXPECTED.TXT
@@ -1,5 +1,5 @@
 ----------------------
-Hello! I'm Reo.
+Hello! I'm reo.Reo.
 What can I do for you?
 ----------------------
 ----------------------
